Extracted files from utils and created seperate folders, hooks, contexts, providers so the folder structure is more readable
