+++ src/components/UserOptions.jsx
@@ -1,4 +1,4 @@
-import { ThemeContext } from '../utils/ThemeContext';
+import { ThemeContext } from '../contexts/ThemeContext';
 import { useContext } from 'react';
 function UserOptions({
   pair,
