+++ src/hooks/useSocket.js
@@ -1,4 +1,4 @@
-import { SocketContext } from './SocketProvider';
+import { SocketContext } from '../providers/SocketProvider';
 import { useContext } from 'react';
 
 export const useSocket = () => {
